Welcome to V2 of COMUKOL: Modern UI

+++ src/views/SignIn.vue
@@ -1,43 +1,38 @@
 <template>
-  <div id="signIn" class="pt-1">
+  <div id="signIn" class="pt-1 font-sans">
     <div id="nav">
       <router-link to="/">Login</router-link> |
       <router-link to="/register">Register</router-link>
     </div>
-    <div class="max-w-sm lg:max-w-md rounded overflow-hidden shadow-lg mx-auto">
-      <img class="w-full" src="@/assets/img/join.png" alt="Join">
+    <div class="max-w-sm lg:max-w-md rounded overflow-hidden mx-auto">
       <div class="px-6 py-4 lg:py-2">
-        <div class="font-bold text-xl mb-2">ComuKOL login</div>
-        <form method="post">
+        <form>
           <div v-if="errors.length" class="text-xs">
             <b class="text-blue-700">Please correct the following error(s):</b>
             <ul>
               <li v-for="error in errors" class="text-red-500">{{ error }}<fa-icon :icon="['fas', 'exclamation-circle']" size="1x" class="ml-1"/></li>
             </ul>
           </div>
-        <label class="flex flex-row">
-          <fa-icon :icon="['fas', 'envelope']" size="1x" class="self-center"/>
-          <input v-model="email" type="email" name="email" placeholder="johndoe@gmail.com" class="input mx-auto py-1 lg:py-2 px-2" autocomplete="on">
+        <label class="flex flex-row border-b border-indigo-400 my-8">
+          <fa-icon :icon="['fas', 'envelope']" size="1x" class="self-center" color="indigo"/>
+          <input v-model="email" type="email" name="email" placeholder="johndoe@gmail.com" class="w-full pl-8 border-none mx-auto py-1 lg:py-2 px-2 focus:outline-none" autocomplete="on">
         </label>
-        <label class="flex flex-row">
-          <fa-icon  :icon="['fas', 'key']" class="self-center" size="1x"/>
-          <input v-model="password" type="password" name="password" class="input mx-auto py-1 lg:py-2 px-2" autocomplete="off">
+        <label class="flex flex-row border-b border-indigo-400 my-8">
+          <fa-icon  :icon="['fas', 'lock']" class="self-center" size="1x" color="indigo"/>
+          <input v-model="password" type="password" name="password" class="w-full pl-8 border-none mx-auto py-1 lg:py-2 px-2 focus:outline-none" autocomplete="off">
         </label>
-        <input @click.submit.prevent="checkForm()" class="btn float-right border border-green-600 bg-white text-green-600 mr-8 lg:mr-10 p-1 rounded-md hover:bg-green-600 hover:text-white" type="submit" name="login" value="sign in">
+        <input @click.submit.prevent="checkForm()" class="w-full p-2 bg-green-500 text-white cursor-pointer hover:bg-green-700 focus:outline-none" type="submit" name="login" value="sign in">
         </form>
-        <router-link to="/forgotpassword" class="w-full"><p class="text-center text-sm text-red-600 underline">forgot password?</p></router-link>
+        <router-link to="/forgotpassword" class="w-full"><p class="text-center text-sm text-red-600 underline my-6 ytr8">forgot password?</p></router-link>
       </div>
-      <div class="px-6 py-4 lg:py-2 w-full mt-2">
-        <span class="inline-block bg-gray-200 rounded-full px-3 py-1 text-sm font-semibold text-gray-700 mr-2">#respond</span>
-        <span class="inline-block bg-gray-200 rounded-full px-3 py-1 text-sm font-semibold text-gray-700 mr-2">#coordinate</span>
-        <span class="inline-block bg-gray-200 rounded-full px-3 py-1 text-sm font-semibold text-gray-700 mt-1">#eradicate</span>
-      </div>
+      <Modal :msg="res" v-if="showModal"/>
+      <Spinner v-if="showSpinner"/>
     </div>
   </div>
 </template>
 
 <script>
-import router from "../router"
+import {bus} from '@/main.js'
 
 export default {
   name: 'SignIn',
@@ -45,24 +40,34 @@ export default {
     return {
       errors: [],
       email:'',
-      password:''
+      password:'',
+      res:{},
+      showModal:false,
+      showSpinner:false
     }
   },
   methods:{
+    modal(message, state){
+      this.showModal = true
+      this.res.text = message
+      this.res.status = state
+    },
     async login(){
+      this.showSpinner = true
       let data={
         email:this.email,
         password:this.password
       }
       let response = await this.$axios.post('/api/login', data)
         .then((response)=>{
-          console.log('Login attempt successful!')
+          this.showSpinner = false
           localStorage.clear()
           localStorage.setItem('token', JSON.stringify(response.data))
-          router.push('/user/profile')
+          this.$router.push('/user/profile')
         })
         .catch((errors)=>{
-          window.alert('Login attempt failed!')
+          this.showSpinner = false
+          this.modal('Wrong email or password',false)
         })
     },
     checkForm(){
@@ -83,6 +88,11 @@ export default {
       let re = /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/
       return re.test(email)
     }
+  },
+  created(){
+    bus.$on('closeModal',(data)=>{
+      this.showModal = data
+    })
   }
 }
 </script>
@@ -101,7 +111,6 @@ export default {
 }
 
 #nav a {
-  font-weight: bold;
   color: #2c3e50;
 }
